Cleanup: Put `Incrementor` helper class in a separate module
Rather than defining this helper class directly in the `platform` branch
module, wrap it in a separate leaf module. (But keep it in the same file
for simplicity.)

This seems more elegant; and it avoids redundant conditionals.
