+++ src/platform/inprocess/mod.rs
@@ -21,7 +21,7 @@ use std::mem;
 use std::usize;
 use uuid::Uuid;
 
-use super::Incrementor;
+use super::incrementor::Incrementor;
 
 #[derive(Clone)]
 struct ServerRecord {
