+++ src/platform/unix/mod.rs
@@ -30,7 +30,7 @@ use std::thread;
 use mio::unix::EventedFd;
 use mio::{Poll, Token, Events, Ready, PollOpt};
 
-use super::Incrementor;
+use super::incrementor::Incrementor;
 
 const MAX_FDS_IN_CMSG: u32 = 64;
